# DocsDev/agent/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LojaSync",
    "version": "1.2.3"
  },
  "paths": {
...
    "/actions/parser-grades": {
      "post": {
        "summary": "Start Grade Parser",
        "operationId": "start_grade_parser_actions_parser_grades_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_start_grade_parser_actions_parser_grades_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GradeExtractionStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/actions/parser-grades/status/{job_id}": {
      "get": {
        "summary": "Parser Grades Status",
        "operationId": "parser_grades_status_actions_parser_grades_status__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GradeExtractionStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Parser Grades Cleanup",
        "operationId": "parser_grades_cleanup_actions_parser_grades_status__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Parser Grades Cleanup Actions Parser Grades Status  Job Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/actions/parser-grades/result/{job_id}": {
      "get": {
        "summary": "Parser Grades Result",
        "operationId": "parser_grades_result_actions_parser_grades_result__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GradeExtractionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/actions/import-romaneio": {
      "post": {
        "summary": "Import Romaneio",
        "operationId": "import_romaneio_actions_import_romaneio_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_romaneio_actions_import_romaneio_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportRomaneioStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/actions/import-romaneio/status/{job_id}": {
      "get": {
        "summary": "Import Romaneio Status",
        "operationId": "import_romaneio_status_actions_import_romaneio_status__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportRomaneioStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Import Romaneio Cleanup",
        "operationId": "import_romaneio_cleanup_actions_import_romaneio_status__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Import Romaneio Cleanup Actions Import Romaneio Status  Job Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/actions/import-romaneio/result/{job_id}": {
      "get": {
        "summary": "Import Romaneio Result",
        "operationId": "import_romaneio_result_actions_import_romaneio_result__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportRomaneioResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_import_romaneio_actions_import_romaneio_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_romaneio_actions_import_romaneio_post"
      },
...
      "Body_start_grade_parser_actions_parser_grades_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_start_grade_parser_actions_parser_grades_post"
      },
...
      "GradeExtractionProduct": {
        "properties": {
          "codigo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codigo"
          },
          "nome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nome"
          },
          "grades": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Grades"
          },
          "atualizado": {
            "type": "boolean",
            "title": "Atualizado"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "atualizado"
        ],
        "title": "GradeExtractionProduct"
      },
      "GradeExtractionResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "total_itens": {
            "type": "integer",
            "title": "Total Itens"
          },
          "total_atualizados": {
            "type": "integer",
            "title": "Total Atualizados"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "itens": {
            "items": {
              "$ref": "#/components/schemas/GradeExtractionProduct"
            },
            "type": "array",
            "title": "Itens"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "status",
          "total_itens",
          "total_atualizados"
        ],
        "title": "GradeExtractionResponse"
      },
      "GradeExtractionStartResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          }
        },
        "type": "object",
        "required": [
          "job_id"
        ],
        "title": "GradeExtractionStartResponse"
      },
      "GradeExtractionStatusResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "started_at": {
            "type": "number",
            "title": "Started At"
          },
          "updated_at": {
            "type": "number",
            "title": "Updated At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "stage",
          "message",
          "started_at",
          "updated_at"
        ],
        "title": "GradeExtractionStatusResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportRomaneioResultResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "saved_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Saved File"
          },
          "local_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local File"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "total_itens": {
            "type": "integer",
            "title": "Total Itens",
            "default": 0
          },
          "grades_disponiveis": {
            "type": "boolean",
            "title": "Grades Disponiveis",
            "default": false
          },
          "total_grades_disponiveis": {
            "type": "integer",
            "title": "Total Grades Disponiveis",
            "default": 0
          },
          "imported_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Imported Keys"
          },
          "import_batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Import Batch Id"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ImportRomaneioResultResponse"
      },
      "ImportRomaneioStartResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          }
        },
        "type": "object",
        "required": [
          "job_id"
        ],
        "title": "ImportRomaneioStartResponse"
      },
      "ImportRomaneioStatusResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "started_at": {
            "type": "number",
            "title": "Started At"
          },
          "updated_at": {
            "type": "number",
            "title": "Updated At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "stage",
          "message",
          "started_at",
          "updated_at"
        ],
        "title": "ImportRomaneioStatusResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}